Redesign for better integration with powsybl (#115)
