+++ src/main/java/org/gridsuite/modification/server/dto/LoadCreationInfos.java
@@ -6,6 +6,9 @@
  */
 package org.gridsuite.modification.server.dto;
 
+import com.powsybl.commons.reporter.Reporter;
+import com.powsybl.commons.reporter.ReporterModel;
+import com.powsybl.iidm.modification.NetworkModification;
 import com.powsybl.iidm.network.LoadType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
@@ -13,10 +16,10 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
+import org.gridsuite.modification.server.NetworkModificationException;
 import org.gridsuite.modification.server.entities.equipment.creation.EquipmentCreationEntity;
 import org.gridsuite.modification.server.entities.equipment.creation.LoadCreationEntity;
-import org.gridsuite.modification.server.modifications.AbstractModification;
-import org.gridsuite.modification.server.modifications.GeneratorCreation;
+import org.gridsuite.modification.server.modifications.LoadCreation;
 
 /**
  * @author Franck Lecuyer <franck.lecuyer at rte-france.com>
@@ -50,7 +53,18 @@ public class LoadCreationInfos extends InjectionCreationInfos {
     }
 
     @Override
-    public AbstractModification toModification() {
-        return new GeneratorCreation(this);
+    public NetworkModification toModification() {
+        return new LoadCreation(this);
+    }
+
+    @Override
+    public NetworkModificationException.Type getErrorType() {
+        return NetworkModificationException.Type.CREATE_LOAD_ERROR;
+    }
+
+    @Override
+    public Reporter createSubReporter(ReporterModel reporter) {
+        String subReportId = "Load creation " + getEquipmentId();
+        return reporter.createSubReporter(subReportId, subReportId);
     }
 }
