+++ src/main/java/org/gridsuite/modification/server/entities/ModificationEntity.java
@@ -69,7 +69,6 @@ public class ModificationEntity {
     }
 
     public void update(ModificationInfos modificationInfos) {
-        return;
     }
 
     //From https://vladmihalcea.com/the-best-way-to-map-a-onetomany-association-with-jpa-and-hibernate/
