+++ src/main/java/org/gridsuite/modification/server/modifications/ModificationUtils.java
@@ -1,0 +1,89 @@
+/**
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.gridsuite.modification.server.modifications;
+
+import com.powsybl.iidm.network.*;
+import org.gridsuite.modification.server.NetworkModificationException;
+
+import static org.gridsuite.modification.server.NetworkModificationException.Type.*;
+
+/**
+ * @author Slimane Amar <slimane.amar at rte-france.com>
+ */
+// TODO transfer to powsybl-core (com.powsybl.iidm.modification)
+final class ModificationUtils {
+
+    private ModificationUtils() {
+    }
+
+    static ModificationUtils getInstance() {
+        return new ModificationUtils();
+    }
+
+    VoltageLevel getVoltageLevel(Network network, String voltageLevelId) {
+        VoltageLevel voltageLevel = network.getVoltageLevel(voltageLevelId);
+        if (voltageLevel == null) {
+            throw new NetworkModificationException(VOLTAGE_LEVEL_NOT_FOUND, voltageLevelId);
+        }
+        return voltageLevel;
+    }
+
+    Bus getBusBreakerBus(VoltageLevel voltageLevel, String busId) {
+        VoltageLevel.BusBreakerView busBreakerView = voltageLevel.getBusBreakerView();
+        Bus bus = busBreakerView.getBus(busId);
+        if (bus == null) {
+            throw new NetworkModificationException(BUS_NOT_FOUND, busId);
+        }
+        return bus;
+    }
+
+    int createNodeBreakerCellSwitches(VoltageLevel voltageLevel, String busBarSectionId, String equipmentId,
+                                      String equipmentName) {
+        return createNodeBreakerCellSwitches(voltageLevel, busBarSectionId, equipmentId, equipmentName, "");
+    }
+
+    int createNodeBreakerCellSwitches(VoltageLevel voltageLevel, String busBarSectionId, String equipmentId,
+                                      String equipmentName, String sideSuffix) {
+        VoltageLevel.NodeBreakerView nodeBreakerView = voltageLevel.getNodeBreakerView();
+        BusbarSection busbarSection = nodeBreakerView.getBusbarSection(busBarSectionId);
+        if (busbarSection == null) {
+            throw new NetworkModificationException(BUSBAR_SECTION_NOT_FOUND, busBarSectionId);
+        }
+
+        // creating the disconnector
+        int newNode = nodeBreakerView.getMaximumNodeIndex();
+        String disconnectorId = "disconnector_" + equipmentId + sideSuffix;
+        String disconnectorName = equipmentName != null ? "disconnector_" + equipmentName + sideSuffix : null;
+        nodeBreakerView.newSwitch()
+            .setId(disconnectorId)
+            .setName(disconnectorName)
+            .setKind(SwitchKind.DISCONNECTOR)
+            .setRetained(false)
+            .setOpen(false)
+            .setFictitious(false)
+            .setNode1(busbarSection.getTerminal().getNodeBreakerView().getNode())
+            .setNode2(newNode + 1)
+            .add();
+
+        // creating the breaker
+        String breakerId = "breaker_" + equipmentId + sideSuffix;
+        String breakerName = equipmentName != null ? "breaker_" + equipmentName + sideSuffix : null;
+        nodeBreakerView.newSwitch()
+            .setId(breakerId)
+            .setName(breakerName)
+            .setKind(SwitchKind.BREAKER)
+            .setRetained(false)
+            .setOpen(false)
+            .setFictitious(false)
+            .setNode1(newNode + 1)
+            .setNode2(newNode + 2)
+            .add();
+
+        return newNode + 2;
+    }
+}
+
